Separate rpc Ethereum context into its file
